+++ server/bulogic.hpp
@@ -47,7 +47,8 @@ namespace BusinessLogic
 
     void balance_stats(Database &database, Session &session, Request &request,
             std::vector<Category::BalanceData> &data);
-
+    void daily_overview(Database &database, Session &session, Request &request,
+            std::vector<std::pair<Category, Category::BalanceData>> &data);
 }
 
 #endif